stat-pusher/requirements.txt
prometheus_client
requests
urllib
python_dateutil
datetime
